Move ServerManager to common module
